% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getreturns.R
\name{getreturns}
\alias{getreturns}
\title{Create a long data frame with log or normal returns.}
\usage{
getreturns(tickers = tickers, from = from, to = to, type = type)
}
\arguments{
\item{tickers}{A long dataframe with columns 'date', 'series', and 'value'.}

\item{from}{Start Date}

\item{to}{End Date}

\item{returntype}{"log" versus "normal"}
}
\value{
A long dataframe of log or normal returns
}
\description{
This function takes a vector of stock tickers and a date range to
generates a dataframe of returns.
}
