% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multireg.R
\name{multiRegAuto}
\alias{multiRegAuto}
\title{Creates recursive multivariate regressions for wide data frames}
\usage{
multiRegAuto(data)
}
\arguments{
\item{data}{A wide dataframe as input. First column must be date, second must be dependent variable,
all columns thereafter are independent variable.}
}
\description{
This function takes a wide dataframe as input and generates recursive multivariate
regressions. First column must be date, second must be dependent variable,
all columns thereafter are independent variable.
}
