% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollcor.R
\name{rollcor}
\alias{rollcor}
\title{Create a plot of rolling correlations of returns for 3 different time frames.}
\usage{
rollcor(data, tf1, tf2, tf3)
}
\arguments{
\item{data}{A wide dataframe with columns 'date', 'returns1', 'returns2'. In that order.}

\item{tf1}{Time frame 1}

\item{tf2}{time frame 2}

\item{tf3}{Time frame 3}
}
\value{
An interactive Plotly plot.
}
\description{
This function takes a wide dataframe as input and generates a plot of rolling correlations
using the ggplot library.
}
